Hajju - 유저가 모임장인 모임이 있으면 탈퇴 못함

+++ src/main/java/com/green/gajigaji/user/UserService.java
@@ -29,7 +29,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
 
 import static com.green.gajigaji.common.GlobalConst.*;
 import static com.green.gajigaji.user.usercommon.UserErrorMessage.*;
@@ -190,11 +189,14 @@ public class UserService {
         if (!userRepository.existsUserEntityByUserSeq(userPk)) {
             throw new CustomException(NOT_FOUND_MESSAGE);
         }
+        if (mapper.existingPartiesCreatedByUser(userPk) != 0) {
+            throw new CustomException(USER_PARTIES_EXIST_ERROR);
+        }
+
         try {
             String midPath = String.format("user/%d", userPk);
             String delAbsoluteFolderPath = String.format("%s%s", customFileUtils.uploadPath, midPath);
             customFileUtils.deleteFolder(delAbsoluteFolderPath);
-
         } catch (Exception e) {
             throw new CustomException(CommonErrorCode.INTERNAL_SERVER_ERROR);
         }
